[3/6] Arm(R) Ethos(TM)-U NPU TIR compiler with conv2d support (#8806)
* Arm(R) Ethos(TM)-U NPU TIR compiler with conv2d support

This commit adds the lowering passes necessary to lower
an NPU Relay module down to a TIR module that can be
compiled for the NPU. Conv2d is supported as the first
NPU operator. An intermediate TE stage between Relay and
TIR allows support for scheduling the operators.


* Fix Conv2D TIR type sensitivity


* Arm(R) Ethos(TM)-U NPU TIR passes and TE for Conv2D

*fixing tests


* Fix import guards for tests


* Fix typing failures with ignores


* Remove unused import


* Reintroduce get_target_accel_type

+++ python/tvm/relay/backend/contrib/ethosu/te/convolution.py
@@ -140,7 +140,7 @@ def conv2d_compute(
             ).astype(ifm.dtype)
             * weight[cc, rh, rw, rc].astype(ifm.dtype)
             # This is a trick to load 10 elements of the scale_bias at once, not accurate maths
-            + (scale_bias[cc, 0] * scale_bias[cc, 9]),
+            + (scale_bias[cc, 0] * scale_bias[cc, 9]).astype(ifm.dtype),
             axis=[rh, rw, rc],
         ),
         name="ethosu_conv2d",
